requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==4.2.26
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-redis==6.0.0
djangorestframework==3.16.1
drf-yasg==1.21.11
Faker==37.12.0
gunicorn==23.0.0
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
profanity==1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
